refactor: move Zod schemas to centralized types.ts
- Create types.ts with TaskExampleSchema and SearchResultSchema
- Export inferred types from Zod schemas for consistency
- Update imports in storage-interface.ts and chroma.ts
- Maintain Zod-first approach for type definitions
